BE/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
langchain-google-genai==1.0.8
langchain-core>=0.2.17,<0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cloudinary==1.36.0
email-validator==2.1.0
argon2-cffi==23.1.0
redis==5.0.1
PyJWT==2.8.0